Add files via upload
This code is used in an AWS Lambda. It fetches Tweets from the Twitter search API and writes them into a JSON file in an AWS S3 bucket. To use this code in AWS Lambda, the python packages "boto3", "requests", "unicodedata" and "pandas" have to be added as a layer.