Add files via upload
This text file contains code used in Windows PowerShell to split large files by number of lines. In this case 800'000 lines (800'000 tweets) per file. The "$" at the beginning of every line and at the very end of the code is used to run the whole code in one go instead of line by line. This is very useful when working with large files, as some of the lines will take up much more time to run than others.