Add files via upload
This text file contains code to split files with Windows PowerShell. It splits the input file into smaller files with a maximum of 800'000 lines each. This code was used to split Twitter data, with one tweet per line. The "$" at the beginning of every line of code and at the very end allows you to run the code as a whole instead of line by line. This is especially useful with very large files, where some lines of the code (reading in the file, looping through it for the splitting process) take up much more runtime than others (defining variables).