Add files via upload
This code is used to clean and filter collected Twitter data. It takes files containing raw Twitter data and filters for english tweets, cleans out redundant data, filters according to keywords and saves them into new files. This is in preparation for creating a count of tweets containing specific keywords.