Add files via upload
This code is used to merge the files created with Twitter_Cleaning_And_Filtering code. It merges the files into daily files and comes directly before the last step, the merging of the files over the whole time period and creating the count, done with the code in Twitter_Merging_Total_And_Creating_Count.